requirements.txt
discord.py
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
